Switch json schema producer to use openapi3 library type
